Compares the rates of two variables against one another.

Compares the rates of two variables against one another.

RDS.compare.proportions(first.interval, second.interval, M = 10000)

Arguments

first.interval
An rds.interval.estimate object fit with either "Gile" or "Salganik" uncertainty.

second.interval
An rds.interval.estimate object fit with either "Gile" or "Salganik" uncertainty.

M
The number of bootstrap resamplings to use

Details

This function preforms a bootstrap test comparing the the rates of two variables against one another.

Examples

if (FALSE) {
data(faux)
int1 <- RDS.bootstrap.intervals(faux, outcome.variable=c("X"), 
weight.type="RDS-II", uncertainty="Salganik", N=1000,
number.ss.samples.per.iteration=1000, 
  confidence.level=0.95, number.of.bootstrap.samples=100)
int2 <- RDS.bootstrap.intervals(faux, outcome.variable=c("Y"), 
  weight.type="RDS-II", uncertainty="Salganik", N=1000,
number.ss.samples.per.iteration=1000,
confidence.level=0.95, number.of.bootstrap.samples=100)
RDS.compare.proportions(int1,int2)
}