Compares the rates of two variables against one another.

Compares the rates of two variables against one another.

RDS.compare.two.proportions(
  data,
  variables,
  confidence.level = 0.95,
  number.of.bootstrap.samples = 5000,
  plot = FALSE,
  seed = 1
)

Arguments

data
An object of class rds.interval.estimates.list with attribute
variables containing a character vector of names of objects of class
rds.interval.estimate.

variables
A character vector of column names to select from data.

confidence.level
The confidence level for the confidence intervals. The default is 0.95 for 95%.

number.of.bootstrap.samples
The number of Monte Carlo draws to determine the null distribution of the likelihood ratio statistic.

plot
Logical, if TRUE then a plot is produces of the null distribution of the likelihood ratio statistic with the observed statistics plotted as a vertical dashed line.

seed
The value of the random number seed. Preset by default to allow reproducability.

Value

An object of class pvalue.table containing the cross-tabulation of p-values for comparing the two classes