A Simulated RDS Data Set with extreme seed dependency

This is a faux set used to demonstrate RDS functions and analysis. The population had N=715 nodes. In this case, the sample size is 500 so that there is a relatively large sample fraction (70%). There is homophily on disease status (R=5) and there is differential activity by disease status whereby the infected nodes have mean degree twice that of the uninfected (w=1.8).

Format

An rds.data.frame plus the original network as a network object

Details

In the sampling the seeds are chosen randomly from the infected population, so there is extreme dependency induced by seed selection.

Each sample member is given 2 uniquely identified coupons to distribute to other members of the target population in their acquaintance. Further each respondent distributes their coupons completely at random from among those they are connected to.

With 70% sample, the VH is substantially biased, so the SS (and presumably MA) do much better. We expect the MA to perform a bit better than the SS.

It is network 702 and its sample from YesYes on mosix. Look for "extract702.R"
The original network is included as fauxsycamore.network as a network object.
The data set also includes the data.frame of the RDS data set as fauxsycamore.
Use data(package="RDS") to get a full list of datasets.

References

Gile, Krista J., Handcock, Mark S., 2009. Respondent-driven Sampling: An Assessment of Current Methodology, Sociological Methodology, 40, 285-327.

See also