homophily configuration graph weights

homophily configuration graph weights

hcg.weights(
  rds.data,
  outcome.variable,
  N = NULL,
  small.fraction = FALSE,
  reltol = sqrt(.Machine$double.eps),
  max.optim = 500,
  theta.start = NULL,
  ...
)

Arguments

rds.data
An rds.data.frame

outcome.variable
The variable used to base the weights on.

N
Population size

small.fraction
should a small sample fraction be assumed

reltol
Relative convergence tolerance for the HCG estimator. The algorithm stops if it is unable to reduce the log-likelihood by a factor of reltol * (abs(log-likelihood) + reltol) at a step. Defaults to sqrt(.Machine$double.eps), typically about 1e-8.

max.optim
The number of iterations on the likelihood optimization for the HCG estimator.

theta.start
The initial value of theta used in the likelihood optimization for the HCG estimator. If NULL, the default, it is the margin of the table of counts for the transitions.

...
Unused

Examples

data(fauxtime)
hcg.weights(fauxtime,"var1",N=3000)

#>   [1]  0.000000  0.000000  0.000000 15.770943  0.000000  1.683414 10.409928
#>   [8] 12.297898  5.061994  2.679375  1.740045  2.742382  2.593300  1.009713
#>  [15] 13.667386  1.627727  9.056365  5.061994 15.770943 10.409928  9.161547
#>  [22]  2.532020  3.394489 24.974053  5.234102 24.974053  3.151369 12.297898
#>  [29]  1.369770  8.306716  2.218710  5.985426 13.667386  6.651247 24.974053
#>  [36]  4.848038  2.006449  4.088921  7.930233  2.175459  2.742382  2.064759
#>  [43]  3.036897  2.719710  4.334356  6.303990  9.056365  4.994343 10.409928
#>  [50]  1.740045 10.409928 15.298366  1.815497  4.508524  2.369174  3.824348
#>  [57]  2.011793  3.191248 15.298366 12.297898  3.406666  1.675936  7.482511
#>  [64]  2.145015  4.029736  2.317264 17.217736  1.866407  4.434379  1.267835
#>  [71]  1.866407  5.985426  3.813773  3.394489  2.880565  2.076778  2.532020
#>  [78]  1.023615  4.434379 12.297898  6.303990  2.011793  3.521810  1.378381
#>  [85]  4.029736  9.056365  6.651247  5.644439  4.029736  4.508524  1.243503
#>  [92]  5.644439  9.216182  4.508524 20.295525  4.334356  6.303990  5.061994
#>  [99]  5.644439  7.145696  7.482511  3.824348  1.925884  7.343373  4.994343
#> [106]  9.790754  6.303990  2.877822  3.191248  1.740045 24.974053  2.558802
#> [113]  2.966202  3.151369  4.032176  5.302578  1.743832  5.234102  9.056365
#> [120]  4.029736  8.306716  5.391982  3.005144  2.742382  9.056365  2.661833
#> [127]  6.154897  7.145696  5.985426  1.040535  2.124974 12.483817  2.076778
#> [134]  9.790754 10.409928  6.930963  3.908457  9.161547  1.240440  2.877822
#> [141]  4.463673  6.930963  4.029736 10.409928  2.020537 12.297898  2.076778
#> [148]  5.985426  5.482243  7.482511 10.409928 17.217736  2.558802  8.306716
#> [155]  1.194353  5.061994  1.220894  1.376008 10.409928  2.076778 10.409928
#> [162]  8.306716  2.032087  7.482511  1.686224 12.297898  7.930233  4.848038
#> [169]  4.032176  2.032087  9.790754  2.427078  1.204830  1.743832  1.080727
#> [176]  1.882687  1.734623 10.409928  5.462915 11.412051 20.295525  4.994343
#> [183]  2.877822  2.364407  4.032176  2.580949  2.877822  4.032176  5.482243
#> [190]  4.334356  9.056365  9.790754  2.145015  9.056365  2.347406  5.825712
#> [197]  4.994343  4.508524  2.076778  1.703252  2.143059  2.011793  2.011793
#> [204]  7.145696  9.056365  4.434379  3.151369  2.317264  5.985426  7.930233
#> [211]  2.654992  1.097209  7.930233  2.877822  7.343373  2.742382  1.228338
#> [218] 10.409928  1.383568  7.930233  5.985426  7.343373  4.848038  6.154897
#> [225]  4.029736  7.482511 13.667386 13.667386  9.790754  2.966202  3.813773
#> [232]  5.644439 19.418929  9.056365  5.644439  0.000000  5.644439  0.000000
#> [239] 29.532665  1.922209  3.908457  1.815497  4.994343  7.812602  3.151369
#> [246]  4.323641  9.790754  7.930233  6.651247  2.966202  2.502454  2.661833
#> [253] 11.412051  4.994343  1.734623  4.463673  6.303990  3.521810  6.651247
#> [260]  5.061994  5.234102  1.438911  3.283043  2.532020  6.930963  7.482511
#> [267]  4.434379  2.966202  8.306716  3.031413 17.217736  2.558802  1.974976
#> [274]  1.740045  4.508524  5.644439  3.191248  5.644439  2.507125  4.848038
#> [281]  1.085385  5.985426  3.263189  6.930963  2.877822  6.303990  3.813773
#> [288]  3.036897  1.763989  1.675936  2.032087  4.994343  3.151369  5.061994
#> [295]  2.725210  9.972268 11.412051  4.994343  1.686224  4.848038  1.627727
#> [302]  5.482243 24.974053 11.412051  5.061994  4.029736  2.532020  4.463673
#> [309]  4.848038  7.482511  2.507125  3.283043  6.303990  7.930233 12.297898
#> [316]  5.234102  9.790754  6.930963  7.482511  2.502454  8.306716  3.031413
#> [323]  2.011793 12.483817  5.644439  3.031413  2.558802  9.056365  1.743832
#> [330]  2.654992  4.088921  2.502454  4.882876  6.651247  4.804045  6.303990
#> [337]  5.482243  1.369770 13.667386 12.297898  6.303990  7.515186 10.557203
#> [344]  3.395074  5.644439  8.306716  9.161547  2.719710 11.412051 13.667386
#> [351]  4.994343  2.880565  3.418013  6.154897  9.056365  7.930233  4.994343
#> [358]  2.518094 39.708403  3.862848 17.217736  3.350106 15.770943  4.848038
#> [365]  6.930963 10.409928  5.985426 24.974053  3.813773  1.627727  3.151369
#> [372] 13.667386  5.061994 11.412051  9.056365 13.852798  6.930963 13.667386
#> [379]  4.666660 10.409928  3.394489 24.974053  5.234102  6.930963 19.418929
#> [386]  8.306716  7.930233  1.826231  2.011793  4.334356  2.567848  4.994343
#> [393]  4.463673  4.804045  5.234102  1.416992 12.297898  6.154897  4.848038
#> [400]  2.212088 19.418929  1.866407 13.852798  4.848038 11.412051 12.297898
#> [407]  2.507125 10.409928  4.994343  2.580949  5.985426  6.154897  4.994343
#> [414]  4.334356 11.412051 17.217736  3.446072  4.029736  1.819662  2.654992
#> [421]  2.725210  2.654992  7.482511  8.306716  2.567848  2.580949  3.191248
#> [428]  3.151369  9.056365  1.194353  4.032176  2.742382  1.709113  3.813773
#> [435]  3.406666  6.651247  4.463673  2.315489  7.515186  5.391982  4.508524
#> [442]  2.145015  7.343373  9.056365  2.461635  2.369174  2.454966  8.306716
#> [449]  3.283043  2.212088 10.409928  1.998487  9.790754  2.212088  2.966202
#> [456]  9.161547  3.263189  3.824348  2.317264  7.343373  3.446072  3.521810
#> [463] 10.409928  5.302578  6.930963  2.593300  9.790754  7.343373  5.061994
#> [470]  7.482511  6.930963  4.434379  7.482511  2.679375  5.985426  7.930233
#> [477]  3.824348  2.454966  6.930963  6.154897  2.317264  2.020537  2.654992
#> [484]  4.994343  3.031413  5.985426  4.029736  2.369174  7.482511  4.029736
#> [491]  3.031413  5.644439  7.145696  6.303990  3.191248  3.521810  7.482511
#> [498]  2.175459  4.994343  5.985426  4.334356  2.283985  3.862848  2.654992
#> [505]  5.061994  4.088921  2.877822  5.644439  2.454966  1.709113  1.815497

fauxtime$NETWORK[c(1,100,40,82,77)] <- NA