Imputes missing degree values

Imputes missing degree values

impute.degree(
  rds.data,
  trait.variable = NULL,
  N = NULL,
  method = c("mean", "quantile"),
  quantile = 0.5,
  recruitment.lower.bound = TRUE,
  round.degree = TRUE
)

Arguments

rds.data
an rds.data.frame

trait.variable
the name of the variable in rds.data to stratify the imputation by

N
population size

method
If mean, the weighted mean value is imputed, otherwize a quantile is used.

quantile
If method is "quantile", this is the quantile that is used. Defaults to median

recruitment.lower.bound
If TRUE, then for each individual, the degree is taken to be the minimum of the number of recruits plus one, and the reported degree

round.degree
Should degrees be integer rounded.

Details

This function imputes degree values using the weighted mean or quantile values of the non-missing degrees. Weights are calcualted using Gile's SS if N is not NULL, or RDS-II if it is. If a trait variable is specified, means and quantile are calculated within the levels of the trait variable

Examples

data(faux)
rds.data <- faux
rds.data$network.size[c(1,2,30,52,81,101,108,111)] <- NA
impute.degree(rds.data)

#>   [1] 15 15 17 32 14 25 43 28 11 10 26 21 28 22 22 18 23 19 29 15 25 24 15 13 39
#>  [26] 26 17 21 16 15 21 23 19 19 14 18 36 17 34 13 17 15 18 13 17 16 20 20 26 17
#>  [51] 21 15 14 24 27 18 20 16 16 25  8 16 34 21 32  9 15 17 19 15 31 15 19 24 23
#>  [76] 18 13 22 18 20 15 24 18 21 14 13 20 22  8 15 22 21 15 19 13 32 21 29 43 15
#> [101] 15 14 12 17 20 16 16 15 27 12 15 14 13 24 19 22 21 14  8 14 14  6  6 25 29
#> [126]  8 18 10  9  8 14 14 23 26  8 11 29 13 18  6 11 19 22 22  9 13 13 14 20 16
#> [151] 10  9 21 16 23 23 15 15 14 29 12 23  7 17 14 16 18 15 14 12  4  4 22 17 10
#> [176] 17 18 28 30 14 22 18 11 22 12 16 12 14 21 23 28 15 18 24  9 11 28 13 16 12
#> [201] 25 13 11 22 18 17 16  5 14 15 17 25 12 15 25 15 12 13 15  9 13 15 38 16 10
#> [226] 14 15 17 24 26 20 17 13 14 21 12 17 15 17 11 14 16 12 23 14 13  7 19 11 13
#> [251] 15 20 16 20 22 17 24 19 16 12 23 24 20  6 13 16 20 20 16 15 14 22 15 22 17
#> [276] 16  9 29 27 15 14 15 20 15 13 15 17 18 15 22 17 13 13 13 11 15  9 15 14 17
#> [301] 14 13 13 14 11 31 17 15 13 15 12  9 16 32 17 10 20 19 10 18 17 22  9 16 18
#> [326] 21  9  8 19 17 12  5 10 13 16 12  9 10 18 23 17 16 19 10  8 11 17 12 18 22
#> [351] 11 11 23 23 16 17 21 23 11 21 16 19  4 17 16  9 16 22 27 12 11 22 15  9 24
#> [376] 27 20 13 11 17 13 10 18 22 18 16 25 11 18

impute.degree(rds.data,trait.variable="X")

#>   [1] 15 16 17 32 14 25 43 28 11 10 26 21 28 22 22 18 23 19 29 15 25 24 15 13 39
#>  [26] 26 17 21 16 16 21 23 19 19 14 18 36 17 34 13 17 15 18 13 17 16 20 20 26 17
#>  [51] 21 15 14 24 27 18 20 16 16 25  8 16 34 21 32  9 15 17 19 15 31 15 19 24 23
#>  [76] 18 13 22 18 20 15 24 18 21 14 13 20 22  8 15 22 21 15 19 13 32 21 29 43 15
#> [101] 15 14 12 17 20 16 16 15 27 12 15 14 13 24 19 22 21 14  8 14 14  6  6 25 29
#> [126]  8 18 10  9  8 14 14 23 26  8 11 29 13 18  6 11 19 22 22  9 13 13 14 20 16
#> [151] 10  9 21 16 23 23 15 15 14 29 12 23  7 17 14 16 18 15 14 12  4  4 22 17 10
#> [176] 17 18 28 30 14 22 18 11 22 12 16 12 14 21 23 28 15 18 24  9 11 28 13 16 12
#> [201] 25 13 11 22 18 17 16  5 14 15 17 25 12 15 25 15 12 13 15  9 13 15 38 16 10
#> [226] 14 15 17 24 26 20 17 13 14 21 12 17 15 17 11 14 16 12 23 14 13  7 19 11 13
#> [251] 15 20 16 20 22 17 24 19 16 12 23 24 20  6 13 16 20 20 16 15 14 22 15 22 17
#> [276] 16  9 29 27 15 14 15 20 15 13 15 17 18 15 22 17 13 13 13 11 15  9 15 14 17
#> [301] 14 13 13 14 11 31 17 15 13 15 12  9 16 32 17 10 20 19 10 18 17 22  9 16 18
#> [326] 21  9  8 19 17 12  5 10 13 16 12  9 10 18 23 17 16 19 10  8 11 17 12 18 22
#> [351] 11 11 23 23 16 17 21 23 11 21 16 19  4 17 16  9 16 22 27 12 11 22 15  9 24
#> [376] 27 20 13 11 17 13 10 18 22 18 16 25 11 18

impute.degree(rds.data,trait.variable="X",method="quantile")

#>   [1] 14 15 17 32 14 25 43 28 11 10 26 21 28 22 22 18 23 19 29 15 25 24 15 13 39
#>  [26] 26 17 21 16 15 21 23 19 19 14 18 36 17 34 13 17 15 18 13 17 16 20 20 26 17
#>  [51] 21 14 14 24 27 18 20 16 16 25  8 16 34 21 32  9 15 17 19 15 31 15 19 24 23
#>  [76] 18 13 22 18 20 14 24 18 21 14 13 20 22  8 15 22 21 15 19 13 32 21 29 43 15
#> [101] 14 14 12 17 20 16 16 14 27 12 14 14 13 24 19 22 21 14  8 14 14  6  6 25 29
#> [126]  8 18 10  9  8 14 14 23 26  8 11 29 13 18  6 11 19 22 22  9 13 13 14 20 16
#> [151] 10  9 21 16 23 23 15 15 14 29 12 23  7 17 14 16 18 15 14 12  4  4 22 17 10
#> [176] 17 18 28 30 14 22 18 11 22 12 16 12 14 21 23 28 15 18 24  9 11 28 13 16 12
#> [201] 25 13 11 22 18 17 16  5 14 15 17 25 12 15 25 15 12 13 15  9 13 15 38 16 10
#> [226] 14 15 17 24 26 20 17 13 14 21 12 17 15 17 11 14 16 12 23 14 13  7 19 11 13
#> [251] 15 20 16 20 22 17 24 19 16 12 23 24 20  6 13 16 20 20 16 15 14 22 15 22 17
#> [276] 16  9 29 27 15 14 15 20 15 13 15 17 18 15 22 17 13 13 13 11 15  9 15 14 17
#> [301] 14 13 13 14 11 31 17 15 13 15 12  9 16 32 17 10 20 19 10 18 17 22  9 16 18
#> [326] 21  9  8 19 17 12  5 10 13 16 12  9 10 18 23 17 16 19 10  8 11 17 12 18 22
#> [351] 11 11 23 23 16 17 21 23 11 21 16 19  4 17 16  9 16 22 27 12 11 22 15  9 24
#> [376] 27 20 13 11 17 13 10 18 22 18 16 25 11 18