An object of class rds.interval.estimate

This function creates an object of class rds.interval.estimate.

rds.interval.estimate(
  estimate,
  outcome.variable,
  weight.type,
  uncertainty,
  weights,
  N = NULL,
  conf.level = 0.95,
  csubset = ""
)

Arguments

estimate
The numerical point estimate of proportion of the trait.variable.

outcome.variable
A string giving the name of the variable in the rds.data that contains a categorical variable to be analyzed.

weight.type
A string giving the type of estimator to use. The options are "Gile's SS", "RDS-I", "RDS-II", "RDS-I (DS)", and "Arithemic Mean". If NULL it defaults to "Gile's SS".

uncertainty
A string giving the type of uncertainty estimator to use. The options are "SRS", "Gile" and "Salganik". This is usually determined by weight.type to be consistent with the estimator's origins. The estimators RDS-I, RDS-I (DS), and RDS-II default to "Salganik", "Arithmetic Mean" defaults to "SRS" and "Gile's SS" defaults to the "Gile" bootstrap.

weights
A numerical vector of sampling weights for the sample, in order of the sample. They should be inversely proportional to the first-order inclusion probabilites, although this is not assessed or inforced.

N
An estimate of the number of members of the population being sampled. If NULL it is read as the pop.size.mid attribute of the rds.data frame. If that is missing it defaults to 1000.

conf.level
The confidence level for the confidence intervals. The default is 0.95 for 95%.

csubset
A character string representing text to add to the output label. Typically this will be the expression used it define the subset of the data used for the estimate.

Value

An object of class rds.interval.estimate is returned. This is a list with components

estimate: The numerical point estimate of proportion of the trait.variable.

interval: A matrix with six columns and one row per category of trait.variable:

point estimate: The HT estimate of the population mean.

95% Lower Bound: Lower 95% confidence bound.

95% Upper Bound: Upper 95% confidence bound.

Design Effect: The design effect of the RDS.

s.e.: Standard error.

n: Count of the number of sample values with that value of the trait.

References

Gile, Krista J., Handcock, Mark S., 2010, Respondent-driven Sampling: An Assessment of Current Methodology. Sociological Methodology 40, 285-327.

Salganik, M., Heckathorn, D. D., 2004. Sampling and estimation in hidden populations using respondent-driven sampling. Sociological Methodology 34, 193-239.

Volz, E., Heckathorn, D., 2008. Probability based estimation theory for Respondent Driven Sampling. The Journal of Official Statistics 24 (1), 79-97.

Author

Mark S. Handcock

RDS.II.estimatesRDS.II.estimates

RDS.SS.estimatesRDS.SS.estimates

Examples

data(faux)
RDS.I.estimates(rds.data=faux,outcome.variable='X',smoothed=TRUE)

#> RDS-I (DS) Estimate for X 
#>      Estimate         95% Interval Design Effect Std. Error     N
#> blue   0.3109 (  0.2681,   0.3537)          1.42     0.0219   118
#> red    0.6891 (  0.6463,   0.7319)          1.42     0.0219   271
#>                                                       Total   389
#> * Using population size estimate: 1000