Create RDS samples with given characteristics

Create RDS samples with given characteristics

rdssampleC(
  net,
  nnodes = network.size(net),
  nsamp0,
  fixinitial,
  nsamp,
  replace,
  coupons,
  select = NULL,
  bias = NULL,
  rds.samp = NULL,
  seed.distribution = NULL,
  attrall = FALSE,
  trait.variable = "disease",
  nsims = 1,
  seeds = NULL,
  prob.network.recall = 1,
  verbose = TRUE
)

Arguments

net
the network object from which to draw a sample

nnodes
the number of nodes in the network [at least as default]

nsamp0
the number of seeds to be drawn (i.e. the size of the 0th wave of sampling)

fixinitial
a variable that indicates the distribution from which to draw the initial seeds, if the seeds variable is NULL and the seed.distribution variable is NULL

nsamp
number of individuals in each RDS sample

replace
sampling with replacement

coupons
number of coupons

select
not used

bias
not used

rds.samp
not used

seed.distribution
a variable [what kind?] that indicates the distribution from which to draw the initial seeds

attrall
Whether all the information about the sample should be returned [??]

trait.variable
attribute of interest

nsims
number of RDS samples to draw

seeds
an array of seeds. Default is NULL, in which case the function draws the seeds from the nodes of the network.

prob.network.recall
simulates the probability that an individual will remember any particular link

verbose
Print verbose output

Value

A list with the following elements: nsample: vector of indices of sampled nodes wsample: vector of waves of each sampled node degsample: vector of degrees of sampled nodes attrsample: vector of attrs of sampled nodes toattr: vector of numbers of referrals to attrsd nodes tonoattr: vector of number of referrans to unattrsd nominators: recruiter of each sample