Plots the recruitment network using the Reingold Tilford algorithm.

Plots the recruitment network using the Reingold Tilford algorithm.

reingold.tilford.plot(
  x,
  vertex.color = NULL,
  vertex.color.scale = hue_pal(),
  vertex.size = 2,
  vertex.size.range = c(1, 5),
  edge.arrow.size = 0,
  vertex.label.cex = 0.2,
  vertex.frame.color = NA,
  vertex.label = get.id(x),
  show.legend = TRUE,
  plot = TRUE,
  ...
)

Arguments

x
An rds.data.frame

vertex.color
The name of the categorical variable in x to color the points with.

vertex.color.scale
The scale to create the color palette.

vertex.size
The size of the vertex points. either a number or the name of a column of x.

vertex.size.range
If vertex.size represents a variable, vertex.size.range is a vector of length 2 representing the minimum and maximum cex for the points.

edge.arrow.size
The size of the arrow from recruiter to recruitee.

vertex.label.cex
The size expansion factor for the vertex.labels.

vertex.frame.color
the color of the outside of the vertex.points.

vertex.label
The name of a variable to use as vertex labels. NA implies no labels.

show.legend
If true and either vertex.color or vertex.size represent variables, legends will be displayed at the bottom of the plot.

plot
Logical, if TRUE then a plot is produced of recruitment tree. ratio statistic with the observed statistics plotted as a vertical dashed line.

...
Additional parameters passed to plot.igraph.

Value

A two-column vector of the positions of the nodes in the recruitment tree.

Examples

if (FALSE) {
data(fauxmadrona)
data(faux)
reingold.tilford.plot(faux)
reingold.tilford.plot(fauxmadrona,vertex.color="disease")
}