Summarizing Generalized Linear Model Fits with Odds Ratios for Survey Data

RDS::summary.svyglm.RDS is a version of summary.svyglm that reports odds-ratios in place of coefficients in the summary table. This only applies for the binomial family. Otherwise it is identical to summary.svyglm. The default in summary.svyglm is to display the log-odds-ratios and this displays the exponetiated from and a 95 p-values are still displayed.

# S3 method for svyglm.RDS
summary(object, correlation = FALSE, df.resid = NULL, odds = TRUE, ...)

Arguments

object
an object of class "svyglm", usually, a result of a call to svyglm.

correlation
logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

df.resid
Optional denominator degrees of freedom for Wald tests.

odds
logical; Should the coefficients be reported as odds (rather than log-odds)?

...
further arguments passed to or from other methods.

Value

RDS::summary.svyglm returns an object of class "summary.svyglm.RDS", a list with components

call

the component from object.

family

the component from object.

deviance

the component from object.

contrasts

the component from object.

df.residual

the component from object.

null.deviance

the component from object.

df.null

the component from object.

deviance.resid

the deviance residuals: see residuals.svyglm.

coefficients

the matrix of coefficients, standard errors, z-values and p-values. Aliased coefficients are omitted.

aliased

named logical vector showing if the original coefficients are aliased.

dispersion

either the supplied argument or the inferred/estimated dispersion if the latter is NULL.

df

a 3-vector of the rank of the model and the number of residual degrees of freedom, plus number of coefficients (including aliased ones).

cov.unscaled

the unscaled (dispersion = 1) estimated covariance matrix of the estimated coefficients.

cov.scaled

ditto, scaled by dispersion.

correlation

(only if correlation is true.) The estimated correlations of the estimated coefficients.

symbolic.cor

(only if correlation is true.) The value of the argument symbolic.cor.

odds

Are the coefficients reported as odds (rather than log-odds)?

Details

svyglm fits a generalised linear model to data from a complex survey design, with inverse-probability weighting and design-based standard errors.

There is no anova method for svyglm as the models are not fitted by maximum likelihood.

See the manual page on svyglm for detail of that function.

See also

Examples

## For examples see example(svyglm)